combined create_cluster and submit_job tests
