+++ dataproc/dataproc_test.go
@@ -1,4 +1,4 @@
-// Copyright 2020 Google LLC
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -28,50 +28,7 @@ import (
 	dataprocpb "google.golang.org/genproto/googleapis/cloud/dataproc/v1"
 )
 
-func setupForSubmitJobTest(projectID, region, clusterName string) error {
-	ctx := context.Background()
-
-	clusterClient, err := dataproc.NewClusterControllerClient(ctx, option.WithEndpoint(fmt.Sprintf("%s-dataproc.googleapis.com:443", region)))
-	if err != nil {
-		return fmt.Errorf("dataproc.NewClusterControllerClient: %v", err)
-	}
-
-	req := &dataprocpb.CreateClusterRequest{
-		ProjectId: projectID,
-		Region:    region,
-		Cluster: &dataprocpb.Cluster{
-			ProjectId:   projectID,
-			ClusterName: clusterName,
-			Config: &dataprocpb.ClusterConfig{
-				MasterConfig: &dataprocpb.InstanceGroupConfig{
-					NumInstances:   1,
-					MachineTypeUri: "n1-standard-1",
-				},
-				WorkerConfig: &dataprocpb.InstanceGroupConfig{
-					NumInstances:   2,
-					MachineTypeUri: "n1-standard-1",
-				},
-			},
-		},
-	}
-
-	// Delete the cluster if it already exists, ignoring any errors.
-	deleteClusterForSubmitJobTest(projectID, region, clusterName)
-
-	// Create the cluster.
-	op, err := clusterClient.CreateCluster(ctx, req)
-	if err != nil {
-		return fmt.Errorf("CreateCluster: %v", err)
-	}
-
-	_, err = op.Wait(ctx)
-	if err != nil {
-		return fmt.Errorf("CreateCluster.Wait: %v", err)
-	}
-	return nil
-}
-
-func deleteClusterForSubmitJobTest(projectID, region, clusterName string) error {
+func deleteCluster(projectID string, clusterName, region string) error {
 	ctx := context.Background()
 
 	clusterClient, err := dataproc.NewClusterControllerClient(ctx, option.WithEndpoint(fmt.Sprintf("%s-dataproc.googleapis.com:443", region)))
@@ -95,24 +52,35 @@ func deleteClusterForSubmitJobTest(projectID, region, clusterName string) error 
 	return nil
 }
 
-func TestSubmitJob(t *testing.T) {
+func TestDataproc(t *testing.T) {
 	tc := testutil.SystemTest(t)
 
-	clusterName := fmt.Sprintf("go-sj-test-%s", tc.ProjectID)
+	clusterName := fmt.Sprintf("go-dp-test-%s", tc.ProjectID)
 	region := "us-central1"
 
-	setupForSubmitJobTest(tc.ProjectID, region, clusterName)
-	defer deleteClusterForSubmitJobTest(tc.ProjectID, region, clusterName)
+	deleteCluster(tc.ProjectID, clusterName, region) // Delete the cluster if it already exists, ignoring any errors.
+	defer deleteCluster(tc.ProjectID, clusterName, region)
 
 	buf := new(bytes.Buffer)
 
 	testutil.Retry(t, 3, 30*time.Second, func(r *testutil.R) {
+		if err := createCluster(buf, tc.ProjectID, region, clusterName); err != nil {
+			r.Errorf("createCluster got err: %v", err)
+			return
+		}
+
+		got := buf.String()
+		if want := fmt.Sprintf("successfully: %s", clusterName); !strings.Contains(got, want) {
+			r.Errorf("CreateCluster: got %s, want %s", got, want)
+			return
+		}
+
 		if err := submitJob(buf, tc.ProjectID, region, clusterName); err != nil {
 			r.Errorf("submitJob got err: %v", err)
 			return
 		}
 
-		got := buf.String()
+		got = buf.String()
 		if want := fmt.Sprint("Job finished successfully"); !strings.Contains(got, want) {
 			r.Errorf("submitJob: got %s, want %s", got, want)
 			return
